requirements.txt
# Versions plus anciennes pour compat macOS (Darwin 16)
numpy==1.21.6
scipy==1.7.3
pandas==1.3.5
scikit-learn==1.0.2
opencv-python==4.5.5.64
torch==1.8.0+cpu
torchvision==0.9.0+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
ultralytics==8.0.20
pillow
requests
